turn namespace functions into Joint methods
