+++ drake/systems/plants/joints/Joint.h
@@ -4,24 +4,10 @@
 #include <random>
 #include "drake/core/Conversions.h"
 #include "drake/util/drakeGeometryUtil.h"
+#include "drake/systems/plants/joints/QuaternionFloating.h"
+#include "drake/systems/plants/joints/RollPitchYawFloating.h"
 
 namespace Drake {
-
-// TODO: move to better place
-template <typename Scalar>
-using Transform3D = Eigen::Transform<Scalar, 3, Eigen::Isometry>;
-
-template <typename Scalar>
-using SpatialVector = Eigen::Matrix<Scalar, TWIST_SIZE, 1>;
-// end move to better place
-
-enum class FloatingBaseType { FIXED = 0, ROLLPITCHYAW = 1, QUATERNION = 2 };
-
-template <typename J>
-class JointType {
- public:
-  virtual ~JointType() {};
-};
 
 template <typename J>
 class Joint {
@@ -41,17 +27,96 @@ class Joint {
   Joint(const std::string &name, const Transform3D<J> &transform_to_parent_body, std::unique_ptr<JointType<J>> type) :
       name(name), transform_to_parent_body(transform_to_parent_body), type(std::move(type)) { };
 
+  template <typename DerivedQ>
+  Transform3D<Promote<J, typename DerivedQ::Scalar>> jointTransform(const Eigen::MatrixBase<DerivedQ> &q) {
+    auto quaternionFloatingJoint = dynamic_cast<const QuaternionFloating<J>*>(type.get());
+    if (quaternionFloatingJoint) {
+      return quaternionFloatingJoint->jointTransform(q);
+    }
+    auto rollPitchYawFloatingJoint = dynamic_cast<const RollPitchYawFloating<J>*>(type.get());
+    if (rollPitchYawFloatingJoint) {
+      return rollPitchYawFloatingJoint->jointTransform(q);
+    }
+    throw std::runtime_error("joint type not handled");
+  }
+
+  template <typename DerivedQ>
+  MotionSubspace<Promote<J, typename DerivedQ::Scalar>> motionSubspace(const Eigen::MatrixBase<DerivedQ> &q) {
+    using Q = typename DerivedQ::Scalar;
+
+    auto quaternionFloatingJoint = dynamic_cast<const QuaternionFloating<J>*>(type.get());
+    if (quaternionFloatingJoint) {
+      return quaternionFloatingJoint->template motionSubspace<Q>();
+    }
+    auto rollPitchYawFloatingJoint = dynamic_cast<const RollPitchYawFloating<J>*>(type.get());
+    if (rollPitchYawFloatingJoint) {
+      return rollPitchYawFloatingJoint->motionSubspace(q);
+    }
+    throw std::runtime_error("joint type not handled");
+  }
+
+  template <typename DerivedQ, typename DerivedV>
+  SpatialVector<Promote<J, typename DerivedQ::Scalar>> motionSubspaceDotTimesV(const Eigen::MatrixBase<DerivedQ> &q, const Eigen::MatrixBase<DerivedV> &v) {
+    using Q = typename DerivedQ::Scalar;
+
+    auto quaternionFloatingJoint = dynamic_cast<const QuaternionFloating<J>*>(type.get());
+    if (quaternionFloatingJoint) {
+      return quaternionFloatingJoint->template motionSubspaceDotTimesV<Q>();
+    }
+    auto rollPitchYawFloatingJoint = dynamic_cast<const RollPitchYawFloating<J>*>(type.get());
+    if (rollPitchYawFloatingJoint) {
+      return rollPitchYawFloatingJoint->motionSubspaceDotTimesV(q);
+    }
+    // TODO
+  };
+
+  template <typename DerivedQ>
+  ConfigurationDerivativeToVelocity<Promote<J, typename DerivedQ::Scalar>> configurationDerivativeToVelocity(const Eigen::MatrixBase<DerivedQ>& q) {
+    using Q = typename DerivedQ::Scalar;
+
+    auto quaternionFloatingJoint = dynamic_cast<const QuaternionFloating<J>*>(type.get());
+    if (quaternionFloatingJoint) {
+      return quaternionFloatingJoint->configurationDerivativeToVelocity(q);
+    }
+    auto rollPitchYawFloatingJoint = dynamic_cast<const RollPitchYawFloating<J>*>(type.get());
+    if (rollPitchYawFloatingJoint) {
+      return rollPitchYawFloatingJoint->template configurationDerivativeToVelocity<Q>();
+    }
+    throw std::runtime_error("joint type not handled");
+  }
+
+  template <typename DerivedQ>
+  VelocityToConfigurationDerivative<Promote<J, typename DerivedQ::Scalar>> velocityToConfigurationDerivative(const Eigen::MatrixBase<DerivedQ>& q) {
+    using Q = typename DerivedQ::Scalar;
+
+    auto quaternionFloatingJoint = dynamic_cast<const QuaternionFloating<J>*>(type.get());
+    if (quaternionFloatingJoint) {
+      return quaternionFloatingJoint->velocityToConfigurationDerivative(q);
+    }
+    auto rollPitchYawFloatingJoint = dynamic_cast<const RollPitchYawFloating<J>*>(type.get());
+    if (rollPitchYawFloatingJoint) {
+      return rollPitchYawFloatingJoint->template velocityToConfigurationDerivative<Q>();
+    }
+    throw std::runtime_error("joint type not handled");
+  }
+
+  template <typename DerivedV>
+  Eigen::Matrix<Promote<J, typename DerivedV::Scalar>, Eigen::Dynamic, 1> frictionTorque(const Eigen::MatrixBase<DerivedV>& v) {
+    using V = typename DerivedV::Scalar;
+
+    auto quaternionFloatingJoint = dynamic_cast<const QuaternionFloating<J>*>(type.get());
+    if (quaternionFloatingJoint) {
+      return quaternionFloatingJoint->template frictionTorque<V>();
+    }
+    auto rollPitchYawFloatingJoint = dynamic_cast<const RollPitchYawFloating<J>*>(type.get());
+    if (rollPitchYawFloatingJoint) {
+      return rollPitchYawFloatingJoint->template frictionTorque<V>();
+    }
+    throw std::runtime_error("joint type not handled");
+  }
+
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW_IF((sizeof(Transform3D<J>)%16)==0)
 };
-
-template <typename Scalar>
-using MotionSubspace = Eigen::Matrix<Scalar, TWIST_SIZE, Eigen::Dynamic, 0, TWIST_SIZE, Joint<Scalar>::MAX_NUM_VELOCITIES>;
-
-template <typename Scalar>
-using ConfigurationDerivativeToVelocity = Eigen::Matrix<Scalar, Eigen::Dynamic, Eigen::Dynamic, 0, Joint<Scalar>::MAX_NUM_VELOCITIES,  Joint<Scalar>::MAX_NUM_POSITIONS>;
-
-template <typename Scalar>
-using VelocityToConfigurationDerivative = Eigen::Matrix<Scalar, Eigen::Dynamic, Eigen::Dynamic, 0, Joint<Scalar>::MAX_NUM_POSITIONS, Joint<Scalar>::MAX_NUM_VELOCITIES>;
 
 }
