+++ drake/systems/plants/joints/test/testJoint.cpp
@@ -1,4 +1,4 @@
-#include "drake/systems/plants/joints/Joints.h"
+#include "drake/systems/plants/joints/Joint.h"
 #include <iostream>
 
 using namespace Drake;
@@ -14,25 +14,12 @@ int main() {
 
   VectorXd q = VectorXd::Random(7);
   cout << j.getName() << endl;
-  cout << jointTransform(j, q).matrix() << endl;
+  cout << j.jointTransform(q).matrix() << endl;
 
   auto q2 = ConvertMatrix<float>()(q);
   auto& q3 = ConvertMatrix<double>()(q);
 
   cout << boolalpha << (&q == &q3) << endl;
 
-
   return 0;
 };
-
-
-
-
-//  template <typename DerivedQ, typename DerivedV>
-//  SpatialVector<Promote<J, typename DerivedQ::Scalar>> motionSubspaceDotTimesV(const Eigen::MatrixBase<DerivedQ> &q, const Eigen::MatrixBase<DerivedV> &v) {
-//    using Q = typename DerivedQ::Scalar;
-//    using V = typename DerivedV::Scalar;
-//    static_assert(std::is_same<Q, V>::value, "types of Q and V must be the same");
-//
-//  };
-
